deleted old logo and dashboard, and moved code
