+++ src/frontend/src/customization/feature-flags.ts
@@ -7,6 +7,4 @@ export const ENABLE_BRANDING = true;
 export const ENABLE_MVPS = false;
 export const ENABLE_CUSTOM_PARAM = false;
 export const ENABLE_INTEGRATIONS = false;
-export const ENABLE_NEW_LOGO = true;
 export const ENABLE_DATASTAX_LANGFLOW = false;
-export const ENABLE_HOMEPAGE = true;
