+++ src/frontend/src/pages/MainPage/components/inputSearchComponent/index.tsx
@@ -1,6 +1,6 @@
+import ForwardedIconComponent from "@/components/common/genericIconComponent";
+import { Input } from "@/components/ui/input";
 import { ChangeEvent, KeyboardEvent } from "react";
-import ForwardedIconComponent from "../../../../../../components/common/genericIconComponent";
-import { Input } from "../../../../../../components/ui/input";
 
 type InputSearchComponentProps = {
   loading: boolean;
