+++ src/frontend/src/pages/MainPage/pages/emptyPage/index.tsx
@@ -1,7 +1,6 @@
 import LangflowLogo from "@/assets/LangflowLogo.svg?react";
 import ForwardedIconComponent from "@/components/common/genericIconComponent";
 import { Button } from "@/components/ui/button";
-import { ENABLE_NEW_LOGO } from "@/customization/feature-flags";
 import { useFolderStore } from "@/stores/foldersStore";
 
 type EmptyPageProps = {
@@ -15,11 +14,7 @@ export const EmptyPage = ({ setOpenModal }: EmptyPageProps) => {
     <div className="m-0 h-full w-full bg-secondary p-0">
       <div className="text-container">
         <div className="relative z-20 flex w-full flex-col items-center justify-center gap-2">
-          {ENABLE_NEW_LOGO ? (
-            <LangflowLogo className="h-7 w-8" />
-          ) : (
-            <span className="fill-foreground text-4xl">⛓️</span>
-          )}
+          <LangflowLogo className="h-7 w-8" />
           <h3
             className="pt-5 font-chivo text-2xl font-semibold text-foreground"
             data-testid="mainpage_title"
